Updated update_and_render_graph.py to render news.
updated update_and_render_graph.py so that every day it will use yf api to look up the top 3 finance news stories of the day. This is a nice info feature as the neural network is best at predicting EOD returns based upon past EOD returns. it has no access to processing live news and adjusting EOD values based upon it. so it's nice to be able to view the days the AI predicted terribly incorrectly to see if there was big news, positive or negative. news is saved in news.csv which will be automatically made when the script runs in the evening.